# openapi.yaml
openapi: 3.1.0
info:
  title: Persistor API
  description: Knowledge graph and vector memory service with tenant isolation, encryption at rest, and real-time notifications.
  version: 0.7.0
  license:
    name: AGPL-3.0
servers:
  - url: http://localhost:3030/api/v1
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key mapped to a single tenant. SHA-256 hashed before storage.
  schemas:
...
    NodeCreate:
      type: object
      required: [type, label]
      properties:
        id:
          type: string
          maxLength: 255
        type:
          type: string
          maxLength: 100
        label:
          type: string
          maxLength: 10000
        properties:
          type: object
...
    Edge:
      type: object
      properties:
        source:
          type: string
          maxLength: 255
        target:
          type: string
          maxLength: 255
        relation:
          type: string
          maxLength: 255
        properties:
          type: object
          description: Encrypted at rest.
        weight:
          type: number
          format: double
          minimum: 0
          maximum: 1000
        salience_score:
          type: number
          format: double
        access_count:
          type: integer
        user_boosted:
          type: boolean
        superseded_by:
          type:
            - string
            - "null"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EdgeCreate:
      type: object
      required: [source, target, relation]
      properties:
        source:
          type: string
          maxLength: 255
        target:
          type: string
          maxLength: 255
        relation:
          type: string
          maxLength: 255
        properties:
          type: object
        weight:
          type: number
          format: double
          minimum: 0
          maximum: 1000
          default: 1.0
    EdgeUpdate:
      type: object
      properties:
        properties:
          type: object
        weight:
          type: number
          format: double
          minimum: 0
          maximum: 1000
    PatchPropertiesRequest:
      type: object
      required: [properties]
      description: >
        Partially update properties with merge semantics. Keys with non-null
        values are added or updated; keys with explicit null values are removed.
      properties:
        properties:
          type: object
          description: Key-value pairs to merge. Null values delete the key.
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
paths:
...
  /edges:
    get:
      summary: List edges
      operationId: listEdges
      tags: [Edges]
      parameters:
        - name: source
          in: query
          schema:
            type: string
        - name: target
          in: query
          schema:
            type: string
        - name: relation
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 1000
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated edge list
          content:
            application/json:
              schema:
                type: object
                properties:
                  edges:
                    type: array
                    items:
                      $ref: "#/components/schemas/Edge"
                  has_more:
                    type: boolean

    post:
      summary: Create an edge
      operationId: createEdge
      tags: [Edges]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EdgeCreate"
      responses:
        "201":
          description: Edge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Edge"
        "400":
          description: Source or target node does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Edge already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /edges/{source}/{target}/{relation}:
    parameters:
      - name: source
        in: path
        required: true
        schema:
          type: string
      - name: target
        in: path
        required: true
        schema:
          type: string
      - name: relation
        in: path
        required: true
        schema:
          type: string

    put:
      summary: Update an edge
      operationId: updateEdge
      tags: [Edges]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EdgeUpdate"
      responses:
        "200":
          description: Edge updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Edge"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete an edge
      operationId: deleteEdge
      tags: [Edges]
      responses:
        "200":
          description: Edge deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                    example: true
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /edges/{source}/{target}/{relation}/properties:
    parameters:
      - name: source
        in: path
        required: true
        schema:
          type: string
      - name: target
        in: path
        required: true
        schema:
          type: string
      - name: relation
        in: path
        required: true
        schema:
          type: string

    patch:
      summary: Partially update edge properties
      description: >
        Merges the supplied properties into the existing edge properties.
        Keys with non-null values are added or updated; keys with explicit
        null values are removed. Other existing properties are left intact.
      operationId: patchEdgeProperties
      tags: [Edges]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchPropertiesRequest"
      responses:
        "200":
          description: Edge with updated properties
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Edge"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /graph/context/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Full context (node + neighbors + edges)
      operationId: graphContext
      tags: [Graph]
      responses:
        "200":
          description: Context bundle
          content:
            application/json:
              schema:
                type: object
...
  /bulk/nodes:
    post:
      summary: Bulk upsert nodes
      description: Accepts up to 1000 nodes. Existing nodes are updated, new ones created.
      operationId: bulkNodes
      tags: [Bulk]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/NodeCreate"
              maxItems: 1000
      responses:
        "200":
          description: Upsert count
          content:
            application/json:
              schema:
                type: object
                properties:
                  upserted:
                    type: integer
  /bulk/edges:
    post:
      summary: Bulk upsert edges
      description: Accepts up to 1000 edges. Existing edges are updated, new ones created.
      operationId: bulkEdges
      tags: [Bulk]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/EdgeCreate"
              maxItems: 1000
      responses:
        "200":
          description: Upsert count
          content:
            application/json:
              schema:
                type: object
                properties:
                  upserted:
                    type: integer
...